Conflicts:
	archived-applications/cxx_benchmark/benchmark.cc
	src/synergia/collective/space_charge_2d_open_hockney.cc
	src/synergia/collective/space_charge_2d_open_hockney.h
	src/synergia/collective/tests/test_space_charge_2d_open_hockney.cc
	src/synergia/collective/tests/test_space_charge_2d_open_hockney_mpi.cc
	src/synergia/lattice/chef_elements.h
	src/synergia/lattice/chef_lattice.cc
	src/synergia/lattice/chef_lattice.h
	src/synergia/lattice/chef_utils.h
	src/synergia/lattice/element_adaptor.cc
	src/synergia/lattice/mad8_adaptors.cc
	src/synergia/lattice/madx_adaptors.cc
	src/synergia/simulation/fast_mapping.h
	src/synergia/simulation/lattice_simulator.cc
	src/synergia/simulation/tests/CMakeLists.txt
	src/synergia/simulation/tests/chef_elements_fixture.h
	src/synergia/simulation/tests/test_propagator.cc

+++ src/synergia/lattice/lattice_element.cc
@@ -318,7 +318,7 @@ Lattice_element::as_string() const
             sstream << ", ";
         }
         sstream << it->first << "=" << "{";
-        for (int i=0; i != (it->second).size(); ++i) {
+        for (size_t i=0; i != (it->second).size(); ++i) {
             if (i) {
                 sstream << ", ";
             }
@@ -329,11 +329,13 @@ Lattice_element::as_string() const
     return sstream.str();
 }
 
+
 void
 Lattice_element::print() const
 {
     std::cout << as_string() << std::endl;
 }
+
 
 template<class Archive>
     void
